requirements.txt
torch
torchvision
monai
transformers
datasets
pandas
numpy
matplotlib
seaborn
pillow
opencv-python
scikit-learn
tqdm
